Add a second try

+++ pkg/reaper/reaper.go
@@ -23,45 +23,52 @@ func Start() {
 }
 
 func killDLV() error {
-	procs, err := procfs.AllProcs()
-	if err != nil {
-		return err
-	}
-
-	var dlvPid int
-	for _, proc := range procs {
-		cl, err := proc.CmdLine()
-		if err != nil {
-			return err
-		}
-		if len(cl) > 2 && cl[0] == "dlv" && cl[1] == "exec" {
-			dlvPid = proc.PID
-			break
-		}
-	}
-
-	if dlvPid == 0 {
-		return nil
-	}
-
-	time.Sleep(8 * time.Second)
-
-	procs, err = procfs.AllProcs()
-	if err != nil {
-		return err
-	}
-
-	for _, proc := range procs {
-		ps, err := proc.Stat()
+	sleep := 8 * time.Second
+	for {
+		procs, err := procfs.AllProcs()
 		if err != nil {
 			return err
 		}
 
-		if ps.PPID == dlvPid {
+		var dlvPid int
+		for _, proc := range procs {
+			cl, err := proc.CmdLine()
+			if err != nil {
+				return err
+			}
+			if len(cl) > 2 && cl[0] == "dlv" && cl[1] == "exec" {
+				dlvPid = proc.PID
+				break
+			}
+		}
+
+		if dlvPid == 0 {
 			return nil
 		}
-	}
 
-	logrus.Infof("Killing dlv: %d", dlvPid)
-	return syscall.Kill(dlvPid, syscall.SIGTERM)
+		time.Sleep(sleep)
+
+		procs, err = procfs.AllProcs()
+		if err != nil {
+			return err
+		}
+
+		for _, proc := range procs {
+			ps, err := proc.Stat()
+			if err != nil {
+				return err
+			}
+
+			if ps.PPID == dlvPid {
+				return nil
+			}
+		}
+
+		if sleep == 0 {
+			logrus.Infof("Killing dlv: %d", dlvPid)
+			return syscall.Kill(dlvPid, syscall.SIGTERM)
+		} else {
+			sleep = 0
+		}
+	}
 }
